pyproject.toml
[tool.poetry]
name = "google_maps_scraper"
version = "0.1.0"
description = ""
authors = ["Jason Fan <jason@finic.ai>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pydantic = "^2.9.1"
playwright = "^1.46.0"
python-dotenv = "^1.0.1"
finicapi = {path = "/Users/jasonfan/Documents/code/finic/python_library", develop = true}

[tool.poetry.scripts]
start = "google_maps_scraper:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
